Move PythonTask to its own file
I think devmate got confused and put its python annotation into mailbox while the implementation is in actor. Lets put it in its own file so it is less confusing.

Differential Revision: [D78921585](https://our.internmc.facebook.com/intern/diff/D78921585/)

[ghstack-poisoned]
