+++ monarch_extension/src/lib.rs
@@ -150,6 +150,11 @@ pub fn mod_init(module: &Bound<'_, PyModule>) -> PyResult<()> {
         "monarch_hyperactor.actor",
     )?)?;
 
+    monarch_hyperactor::tokio::register_python_bindings(&get_or_add_new_module(
+        module,
+        "monarch_hyperactor.tokio",
+    )?)?;
+
     monarch_hyperactor::mailbox::register_python_bindings(&get_or_add_new_module(
         module,
         "monarch_hyperactor.mailbox",
