+++ monarch_hyperactor/src/actor.rs
@@ -7,11 +7,8 @@
  */
 
 use std::error::Error;
-use std::fmt;
-use std::future::Future;
 use std::future::pending;
 use std::ops::Deref;
-use std::pin::Pin;
 use std::sync::Arc;
 use std::sync::OnceLock;
 
@@ -33,7 +30,6 @@ use monarch_types::SerializablePyErr;
 use pyo3::IntoPyObjectExt;
 use pyo3::exceptions::PyBaseException;
 use pyo3::exceptions::PyRuntimeError;
-use pyo3::exceptions::PyStopIteration;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 use pyo3::types::PyBytes;
@@ -60,6 +56,8 @@ use crate::proc::PyProc;
 use crate::proc::PySerialized;
 use crate::runtime::signal_safe_block_on;
 use crate::shape::PyShape;
+use crate::tokio::PyPythonTask;
+use crate::tokio::PythonTask;
 
 #[pyclass(frozen, module = "monarch._rust_bindings.monarch_hyperactor.actor")]
 #[derive(Serialize, Deserialize, Named)]
@@ -673,115 +671,6 @@ impl Handler<PythonMessage> for PythonActor {
     }
 }
 
-/// Helper struct to make a Python future passable in an actor message.
-///
-/// Also so that we don't have to write this massive type signature everywhere
-pub(crate) struct PythonTask {
-    future: Mutex<Pin<Box<dyn Future<Output = PyResult<PyObject>> + Send + 'static>>>,
-}
-
-impl PythonTask {
-    pub(crate) fn new(fut: impl Future<Output = PyResult<PyObject>> + Send + 'static) -> Self {
-        Self {
-            future: Mutex::new(Box::pin(fut)),
-        }
-    }
-
-    fn take(self) -> Pin<Box<dyn Future<Output = PyResult<PyObject>> + Send + 'static>> {
-        self.future.into_inner()
-    }
-}
-
-impl fmt::Debug for PythonTask {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("PythonTask")
-            .field("future", &"<PythonFuture>")
-            .finish()
-    }
-}
-
-#[pyclass(
-    name = "PythonTask",
-    module = "monarch._rust_bindings.monarch_hyperactor.actor"
-)]
-pub struct PyPythonTask {
-    inner: Option<PythonTask>,
-}
-
-impl From<PythonTask> for PyPythonTask {
-    fn from(task: PythonTask) -> Self {
-        Self { inner: Some(task) }
-    }
-}
-
-#[pyclass(
-    name = "JustStopWithValueIterator",
-    module = "monarch._rust_bindings.monarch_hyperactor.actor"
-)]
-struct JustStopWithValueIterator {
-    value: Option<PyObject>,
-}
-
-#[pymethods]
-impl JustStopWithValueIterator {
-    fn __next__(&mut self) -> PyResult<PyObject> {
-        Err(PyStopIteration::new_err(self.value.take().unwrap()))
-    }
-}
-
-impl PyPythonTask {
-    pub fn new<F, T>(fut: F) -> PyResult<Self>
-    where
-        F: Future<Output = PyResult<T>> + Send + 'static,
-        T: for<'py> IntoPyObject<'py>,
-    {
-        Ok(PythonTask::new(async {
-            fut.await
-                .and_then(|t| Python::with_gil(|py| t.into_py_any(py)))
-        })
-        .into())
-    }
-}
-
-#[pymethods]
-impl PyPythonTask {
-    fn into_future(&mut self, py: Python<'_>) -> PyResult<PyObject> {
-        let task = self
-            .inner
-            .take()
-            .map(|task| task.take())
-            .expect("PythonTask already consumed");
-        Ok(pyo3_async_runtimes::tokio::future_into_py(py, task)?.unbind())
-    }
-    fn block_on(&mut self, py: Python<'_>) -> PyResult<PyObject> {
-        let task = self
-            .inner
-            .take()
-            .map(|task| task.take())
-            .expect("PythonTask already consumed");
-        signal_safe_block_on(py, task)?
-    }
-
-    /// In an async context this turns the tokio::Future into
-    /// an asyncio Future and awaits it.
-    /// In a synchronous context, this just blocks on the future and
-    /// immediately returns the value without pausing caller coroutine.
-    /// See [avoiding async code duplication] for justitifcation.
-    fn __await__(&mut self, py: Python<'_>) -> PyResult<PyObject> {
-        let lp = py
-            .import("asyncio.events")
-            .unwrap()
-            .call_method0("_get_running_loop")
-            .unwrap();
-        if lp.is_none() {
-            let value = self.block_on(py)?;
-            Ok(JustStopWithValueIterator { value: Some(value) }.into_py_any(py)?)
-        } else {
-            self.into_future(py)?.call_method0(py, "__await__")
-        }
-    }
-}
-
 async fn handle_async_endpoint_panic(
     panic_sender: UnboundedSender<anyhow::Result<(), SerializablePyErr>>,
     task: PythonTask,
@@ -857,7 +746,6 @@ pub fn register_python_bindings(hyperactor_mod: &Bound<'_, PyModule>) -> PyResul
     hyperactor_mod.add_class::<MethodSpecifier>()?;
     hyperactor_mod.add_class::<UnflattenArg>()?;
     hyperactor_mod.add_class::<PanicFlag>()?;
-    hyperactor_mod.add_class::<PyPythonTask>()?;
     Ok(())
 }
