+++ monarch_hyperactor/src/mailbox.rs
@@ -45,13 +45,13 @@ use pyo3::types::PyType;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::actor::PyPythonTask;
 use crate::actor::PythonMessage;
 use crate::actor::PythonMessageKind;
-use crate::actor::PythonTask;
 use crate::proc::PyActorId;
 use crate::runtime::signal_safe_block_on;
 use crate::shape::PyShape;
+use crate::tokio::PyPythonTask;
+use crate::tokio::PythonTask;
 #[derive(Clone, Debug)]
 #[pyclass(
     name = "Mailbox",
