+++ monarch_rdma/extension/lib.rs
@@ -13,10 +13,10 @@ use hyperactor::Named;
 use hyperactor::ProcId;
 use hyperactor_mesh::RootActorMesh;
 use hyperactor_mesh::shared_cell::SharedCell;
-use monarch_hyperactor::actor::PyPythonTask;
 use monarch_hyperactor::mailbox::PyMailbox;
 use monarch_hyperactor::proc_mesh::PyProcMesh;
 use monarch_hyperactor::runtime::signal_safe_block_on;
+use monarch_hyperactor::tokio::PyPythonTask;
 use monarch_rdma::IbverbsConfig;
 use monarch_rdma::RdmaBuffer;
 use monarch_rdma::RdmaManagerActor;
