+++ python/monarch/_rust_bindings/monarch_hyperactor/alloc.pyi
@@ -10,7 +10,8 @@ from datetime import timedelta
 from typing import final, Optional
 
 from monarch._rust_bindings.monarch_hyperactor.alloc import Alloc, AllocSpec
-from monarch._rust_bindings.monarch_hyperactor.mailbox import Mailbox, PythonTask
+from monarch._rust_bindings.monarch_hyperactor.mailbox import Mailbox
+from monarch._rust_bindings.monarch_hyperactor.tokio import PythonTask
 
 from typing_extensions import Self
